Add internal initializing state to components.
This allows for certain bindings to run on first pass when they might
otherwise not need to. Specifically the if bindings should normally only
run if the node is already in the document, but they still need to run
when internally parsing the template the first time, before the
component has had a chance to mount.