Ensure content restoration is as accurate as possible.
This is done by storing a reference to the nearest sibling nodes in
either direction and inserting either before or after them. If neither
is found, the content should simply be appended normally.

This could be improved further by storing the full list of previous and
subsequent siblings and iterating over each in turn but this will do for
now.